requirements.txt
# Requirements for AI-Driven Learning Platform
requests>=2.28.0
pyjwt>=2.4.0
cryptography>=3.4.8
python-dotenv>=0.19.0
redis>=4.3.0
celery>=5.2.0
async-timeout>=4.0.0
aiofiles>=0.8.0
uvicorn>=0.18.0
fastapi>=0.78.0
pydantic>=1.9.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.5
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
alembic>=1.8.0